--- src/dictionaries/i18n.ts
@@ -1,16 +1,0 @@
-import 'server-only'
-
-const dictionaries: {
-  [key: string]: () => Promise<typeof import('./en.json')>
-} = {
-  en: () => import('./en.json').then((module) => module.default),
-  ko: () => import('./ko.json').then((module) => module.default),
-  zh: () => import('./en.json').then((module) => module.default),
-}
-
-export const getDictionary = async (lang: string) => dictionaries[lang]()
-
-export const i18n = {
-  defaultLang: 'en',
-  langs: ['ko', 'zh'],
-}
